Add files via upload
CNN model architecture with two or more convolution layers, uses ReLU activation, and a pooling layer after each convolution. After the last pooling layer, the images are  flattened and fed into a dense layer with ReLu activation and a final dense layer with Sigmoid activation. This code is run on Google Colab and uses a Histopathology images dataset from Kaggle.